Update open_secrets_api
Took out the printed messages in favor of a cleaner DataFrame. I dig Numpy and Pandas a lot, so I might be doing more work on this front. I was having some problems with PyCharm cutting out some of the dataframe in the console, so I consulted the JetBrains message board and found a quick fix. That is the only non-original piece of code in this program, I swear!